cleanup files. Terminating branch after merge. Moving to Style Dictionary implementation in another branch
